sqlite: Fix the sqlite METADATA template
The metadata template is supposed to generate year, month, and day.
However, the date format characters were wrong.  This uses the correct
format characters and remotes zero-padding, so, for instance, day 1 of
the month is "1" and not "01".

Fixed the 3.44.3 METADATA file by hand.

Tested with a new sqlite release and verified that lint did not detect
errors in the METADATA file.

Flag: EXEMPT bug-fix
Bug: 397436943
Test: presumbit
Change-Id: I2b94a36866d8e5e09837a5f35c509da92ca31f2c